% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ph_ants}
\alias{ph_ants}
\title{Data from: Are ant supercolonies crucibles of a new major transition in evolution?}
\format{
\subsection{\code{ph_ants}}{

A data frame of Procrustes superimposed shape data with 1494 rows and 24 columns:
\describe{
\item{Biosample}{Biosample}
\item{Species}{Species}
...
}
}
}
\source{
\url{http://dx.doi.org/10.5061/dryad.f65bn}
}
\usage{
ph_ants
}
\description{
Abstract: The biological hierarchy of genes, cells, organisms and societies is a fundamental reality in the living world.
This hierarchy of entities did not arise ex nihilo at the origin of life, but rather has been serially generated
by a succession of critical events known as ‘evolutionary transitions in individuality’ (ETIs). Given the sequential
nature of ETIs, it is natural to look for candidates to form the next hierarchical tier. We analyse claims that
these candidates are found among ‘supercolonies’, ant populations in which discrete nests cooperate as part of a
wider collective, in ways redolent of cells in a multicellular organism. Examining earlier empirical work and new
data within the recently proposed ‘Darwinian space’ framework, we offer a novel analysis of the evolutionary status
of supercolonies and show how certain key conditions might be satisfied in any future process transforming these
collaborative networks into true Darwinian individuals.
}
\keyword{datasets}
