% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_equate.R
\name{ph_equate}
\alias{ph_equate}
\title{Equate factors levels across columns.}
\usage{
ph_equate(df, class_col)
}
\arguments{
\item{df}{A \code{data.frame} of column-wise class predictions from each ensemble model.}

\item{class_col}{A \code{character} value for the column name of the actual classes.}
}
\value{
A \code{data.frame} of column-wise class predictions with equivalent class levels.
}
\description{
The \code{ph_equate} function ensures that the factor levels in all columns are equal. When classification are heavily biased or inaccurate, they can return new class predictions that do not contain every level in the original data. This can interfere with model evaluation functions e.g. via a confusion matrix.
}
