% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_eval.R
\name{ph_eval}
\alias{ph_eval}
\title{Evaluate a phenotype classification model.}
\usage{
ph_eval(df, model, class)
}
\arguments{
\item{df}{A \code{data.frame} of unseen data (e.g., validation or test data) that relates to the training data with which \code{model} was trained.}

\item{model}{A \code{train} model.}

\item{class}{A (\code{factor}) value of the actual classes of unseen data.}
}
\value{
A list containing the following components:\tabular{ll}{
\code{preds} \tab The class predictions. \cr
\tab \cr
\code{cm} \tab Evaluation results from the confusion matrix; for the multi-class case, the results are averaged across all class levels. \cr
}
}
\description{
The \code{ph_eval} function generates a confusion matrix for binary or multi-class classification; for the multi-class case, the results are averaged across all class levels.
}
