% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_outs.R
\name{ph_iqr}
\alias{ph_iqr}
\title{Compute interquartile range.}
\usage{
ph_iqr(x, na.rm = FALSE, type = 7)
}
\arguments{
\item{x}{A \code{numeric} vector.}

\item{na.rm}{A \code{logical} value: FALSE (default). If true, any NA is removed before quantiles are computed.}

\item{type}{An \code{integer} value (1:9) selecting one of 9 quantile algorithms.}
}
\value{
The interquartile range.
}
\description{
The \code{ph_iqr} function computes the interquartile range.
}
