% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_outs.R
\name{ph_outs}
\alias{ph_outs}
\title{Find outlier indices.}
\usage{
ph_outs(x)
}
\arguments{
\item{x}{A \code{numeric} vector.}
}
\value{
The outlier indices.
}
\description{
The \code{ph_outs} function computes outliers with the interquartile method.
}
