% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_eval.R
\name{ph_eval}
\alias{ph_eval}
\title{Evaluate a phenotype classification model.}
\usage{
ph_eval(pred, obs)
}
\arguments{
\item{pred}{A \code{factor} value of predicted classes.}

\item{obs}{A \code{factor} value of the observed or actual classes.}
}
\value{
A \code{data.frame} of confusion matrix evaluation results; for the multi-class case, the results are averaged across all class levels.
}
\description{
The \code{ph_eval} function generates a confusion matrix for binary or multi-class classification; for the multi-class case, the results are averaged across all class levels.
}
